pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "playlist-maker"
version = "0.1.0"
description = "Spotify URL -> media download + tagging"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "spotipy",
  "yt-dlp",
  "mutagen",
  "Pillow",
  "requests"
]

[tool.setuptools]
packages = ["spotify2media"]

[project.scripts]
playlist-maker = "spotify2media.cli:run"

